add perf test and pathological case
